feat(#3742): create a separate package for errors
