+++ eo-parser/src/main/java/org/eolang/parser/errors/EoParserErrors.java
@@ -1,0 +1,12 @@
+package org.eolang.parser.errors;
+
+import java.util.Iterator;
+import org.antlr.v4.runtime.BaseErrorListener;
+import org.xembly.Directive;
+
+public final class EoParserErrors extends BaseErrorListener implements Iterable<Directive> {
+    @Override
+    public Iterator<Directive> iterator() {
+        return null;
+    }
+}
