+++ eo-parser/src/test/java/org/eolang/parser/errors/UnderlinedMessageTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.parser;
+package org.eolang.parser.errors;
 
 import java.util.stream.Stream;
 import org.hamcrest.MatcherAssert;
